pyproject.toml
[project]
name = "tbd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "openai>=2.6.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
]
